department added

+++ server_api/src/controllers/collegeController.ts
@@ -73,17 +73,26 @@ const updateCollege = async (req: Request, res: Response): Promise<void> => {
 
 const deleteCollege = async (req: Request, res: Response): Promise<void> => {
   try {
-    const deletedCollege = await College.findByIdAndDelete(req.params.id);
-    if (!deletedCollege) {
+    const college = await College.findById(req.params.id);
+
+    if (!college) {
       errorResponse(res, 404, 'College not found');
       return;
     }
+
+    if (college.number_of_departments > 0) {
+      errorResponse(res, 400, 'Cannot delete college with existing departments');
+      return;
+    }
+
+    await College.findByIdAndDelete(req.params.id);
 
     successResponse(res, 200, 'College deleted successfully', true);
   } catch (error) {
     errorResponse(res, 500, 'Server error', (error as Error).message);
   }
 };
+
 
 const collegeController = {
   createCollege,
